+++ fuel-tx/src/tests/valid_cases/transaction.rs
@@ -53,6 +53,7 @@ fn gas_limit() {
     TransactionBuilder::create(vec![0xfau8].into(), rng.gen(), vec![])
         .maturity(maturity)
         .add_random_fee_input()
+        .add_contract_created()
         .finalize()
         .check(block_height, &test_params())
         .expect("Failed to validate transaction");
@@ -88,6 +89,7 @@ fn maturity() {
     TransactionBuilder::create(rng.gen(), rng.gen(), vec![])
         .maturity(block_height)
         .add_random_fee_input()
+        .add_contract_created()
         .finalize()
         .check(block_height, &test_params())
         .expect("Failed to validate tx create");
@@ -147,6 +149,7 @@ fn create_not_set_witness_limit_success() {
     // When
     let result = TransactionBuilder::create(bytecode.clone().into(), rng.gen(), vec![])
         .add_random_fee_input()
+        .add_contract_created()
         .finalize()
         .check(block_height, &test_params());
 
@@ -208,6 +211,7 @@ fn create_set_witness_limit_for_empty_witness_success() {
     // When
     let result = TransactionBuilder::create(bytecode.clone().into(), rng.gen(), vec![])
         .add_random_fee_input()
+        .add_contract_created()
         .witness_limit(limit as u64)
         .finalize()
         .check(block_height, &test_params());
@@ -282,6 +286,7 @@ fn create_not_set_max_fee_limit_success() {
     // When
     let result = TransactionBuilder::create(rng.gen(), rng.gen(), vec![])
         .add_random_fee_input()
+        .add_contract_created()
         .finalize()
         .check(block_height, &test_params());
 
@@ -349,6 +354,7 @@ fn max_iow() {
         TransactionBuilder::create(generate_bytes(rng).into(), rng.gen(), vec![]);
 
     builder.maturity(maturity);
+    builder.add_contract_created();
 
     let secrets =
         cmp::min(TX_PARAMS.max_inputs() as u32, TX_PARAMS.max_witnesses() - 1) as usize;
@@ -641,6 +647,7 @@ fn create__check__happy_path() {
     TransactionBuilder::create(generate_bytes(rng).into(), rng.gen(), vec![])
         .maturity(maturity)
         .add_random_fee_input()
+        .add_contract_created()
         .finalize()
         .check(block_height, &test_params())
         .expect("Failed to validate tx");
@@ -673,6 +680,26 @@ fn create__check__cannot_have_contract_input() {
     assert_eq!(
         err,
         ValidityError::TransactionInputContainsContract { index: 0 }
+    );
+}
+
+#[test]
+fn create__check__must_contain_contract_created_output() {
+    let rng = &mut StdRng::seed_from_u64(8586);
+
+    let maturity = 100.into();
+    let block_height = 1000.into();
+
+    let err = TransactionBuilder::create(generate_bytes(rng).into(), rng.gen(), vec![])
+        .maturity(maturity)
+        .add_random_fee_input()
+        .finalize()
+        .check(block_height, &test_params())
+        .expect_err("Expected erroneous transaction");
+
+    assert_eq!(
+        err,
+        ValidityError::TransactionOutputDoesntContainContractCreated
     );
 }
 
@@ -716,6 +743,7 @@ fn create__check__cannot_have_variable_output() {
     let err = TransactionBuilder::create(generate_bytes(rng).into(), rng.gen(), vec![])
         .maturity(maturity)
         .add_random_fee_input()
+        .add_contract_created()
         .add_output(Output::variable(rng.gen(), rng.gen(), rng.gen()))
         .finalize()
         .check(block_height, &test_params())
@@ -723,7 +751,7 @@ fn create__check__cannot_have_variable_output() {
 
     assert_eq!(
         err,
-        ValidityError::TransactionOutputContainsVariable { index: 0 }
+        ValidityError::TransactionOutputContainsVariable { index: 1 }
     );
 }
 
@@ -834,6 +862,7 @@ fn create__check__something_else() {
     )
     .maturity(maturity)
     .add_unsigned_coin_input(secret, rng.gen(), rng.gen(), AssetId::default(), rng.gen())
+    .add_contract_created()
     .add_output(Output::change(rng.gen(), rng.gen(), AssetId::default()))
     .finalize()
     .check(block_height, &test_params())
@@ -910,6 +939,7 @@ fn create__check__something() {
             AssetId::default(),
             rng.gen(),
         )
+        .add_contract_created()
         .add_output(Output::change(rng.gen(), rng.gen(), AssetId::default()))
         .finalize()
         .check(block_height, &test_params())
@@ -941,6 +971,7 @@ fn create__check__can_max_out_storage_slots() {
     )
     .maturity(maturity)
     .add_unsigned_coin_input(secret, rng.gen(), rng.gen(), AssetId::default(), rng.gen())
+    .add_contract_created()
     .add_output(Output::change(rng.gen(), rng.gen(), AssetId::default()))
     .finalize()
     .check(block_height, &test_params())